init: adding docker bases for performance study
This is the start of the performance study repository,
where we will store experimental configs and docker build
recipes for running performance experiments across three
clouds: Google, AWS, and Azure. This first commit adds
the early container builds for all of AWS and Google,
for each of CPU and GPU. The difference between AWS and
Google is that AWS builds need libfabric to enable EFA.
We have not yet been able to do any work on Azure as we
do not have resources there. The README describes the
basic organization (TBA). My next stage of work will
be translating final build configurations into the
(currently empty) organized directories for each
of Google and AWS, and then (when possible) Azure.
We did the final configurations on our on-premises
systems, and will update the testing parameters with
these, and then do a final set of runs to test and
time next week.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>