google cloud gke setup for cpu experiments
This adds the skeleton of configs to run
on Google cloud for GKE, with CPU. I am also pinning
the Flux operator version (the crd to install and
the container digest pulled) and setting all the
CRD files to have 32 nodes. We actually only have
one tested config here from on-premises, quicksilver
from Abhik!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>