docker: update builds for zen4
This is a large changeset that ensures that all CPU builds
are done on zen4, via AMD. The previous ones were done on Skylake.
We also better "finalized" the GPU parameters in preparation for
testing. We will not update the other GPU setups until these
are tested. Tomorrow we will get 4 GPU nodes on AWS to do this,
along with testing out AKS to get infiniband drivers working
(at least we hope).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>