eks cpu: testing of apps
Most applications worked - amg2023 did not work on zen4, but
the older intel build did (and I tested that). Akin to AKS,
mixbench-cpu segfaults, mt-gemm needs to have a fix to the
output (it is not meaningful) and kripke parameters are not
done yet. Also not that it is important to completely delete
a minicluster before creating another one - something with
the device can render both in a "stuck state" that while
is recoverable, takes time and thus $$.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>