aws ec2: first run for cpu with singularity!
This is a full setup (image VM builds) for Singularity with flux on ec2.
I am doing this because we needed to test singularity, and Parallel Cluster
has not been working great. I want to have this working setup in case
we need it (and actually I think I prefer it, and believe it to be more
comparable to our other setups that use flux!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>