aks/eks updates
I have rebuilt the mt-gemm base images for CPU (azure and aws)
so that the output is more meaningful - there were parts of the file
that were commented. I have also figured out how to expand the size of
the Azure partition and pull down singularity images to it, saving
a VM image that I will test later today. Finally, I have updated all
of the saving of output to not use tee, and instead get the filename
and output based on the study id. There was also a bug that we were
not capturing stderr with flux job attach.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>