azure vmset: testing cpu on azure with infiniband
This changeset include adding a vmset on Azure that comes with
infiniband and a full set of our experiment containers, automation
to bring up flux, so everything is ready to go. I was not able to
get the proximity group working, but I remember someone said maybe
we should not have it, and I did not see issues with the performance,
but again it was only two nodes.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>