aws eks size256 results and configs
This did not work the first time because we ran out
of addresses. Changing parameters in the aws-cni that
populated the environment along with allowing duplication
seemed to work. But honestly I am not even sure if that
would reproduce - it just suddently started working
after we made these tweaks.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>