compute-engine: google gpu size 8
This includes results exclusing osu all reduce and quicksilver,
which have a bug that seems to be related to GPU registration
that is likely related to using ethernet (we did not see it
on the libfabric or infiniband setups). I am not going to run
these until there is more insight into the error - it is
not a good use of credits.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>