compute-engine gpu size 32
I am going to next prepare a VM base that has flux run
by root, and then build quicksilver and multi-gpu
models on the host (we already have osu). This should
get us the final data that we need. If those builds
work I will test on a size 2, then run for each of
sizes 4,8,16,32 and we will be done with this setup

Signed-off-by: vsoch <vsoch@users.noreply.github.com>