container-sizes: analysis to look at size of layers
This includes work to parse the events files and determine pull times,
along with getting manifests and configs for all unique containers in
the study. We filter this down to those that were used as experiment
apps, and then look at overall layer sizes (histogram) and similarity
based on digests. We also make plots that look at pull times for the
entire study and within applications. This can be expanded if needed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>