gke: gpu to add sizes 4,8,16 for amg2023
The current runs for GKE were just using CPU. This is the start
of work to add runs using a working GPU container. I am still
not able to get the GPU allocation for 32 nodes (gulp) so will
need to keep trying and add that data later. This changeset
also adds the missing amg2023 config for the CycleCloud gpus.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>