chore: aws eks cpu and compute engine gpu
There are a few result files with "unsatisfiable" (which means the runs were done wrong
and then the result saved, and they should have been cleaned up) so I am removing
them. I am also making result files that should not be parsed (e.g., testing)
start with underscore so our analysis scripts know to skip. For this example,
it was a test on 2 nodes before the actual runs that I saved. I do not want
to actually delete them, but I do not want to parse them.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>