lammps: calculate hookup time
The hookup time is important - it is the time that it takes for the job to
start, the networking fabric to hookup, and then the application to run.
The wall time can be really fast, but if the hookup time is really slow,
the actual time you are waiting for a run is much slower than reflected
in the application time. You can either look at the total wrapper time
(the workload manager start to end) or the hookup time separately,
which is wrapper duration - wall time.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>