analysis: osu benchmarks
This adds results for osu_latency, osu_bw, and osu_allreduce.
I did my best to parse these - the data is huge, inconsistent,
and it was very challenging. This should serve as a first
parse to glance at the data and we can further look into the
results. I did my best but a second pair of eyes on my procedure
here would be helpful.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>