analysis: kripke
We do not have data for 32 GPU nodes on AWS - we could
only get up to 16. For size 4 on CycleCloud GPU, we have
results but kripke never finishes to get to the FOMs table.
For Google Cloud GKE size 128, the kripke result is missing.
I can try to re-run it - I did not find it in artifacts and
suspect it either did not finish in enough time to save or
was missed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>