analysis results: stream single node benchmark
This would not work to assemble with pandas - there were too many
data points. Instead I created vectors of values, where each list
was of lists corresponding to environment sizes, and then the
entire set is for an environment. I could then plot them with
matplotlib boxplot and space them (on the same plot) so they
appeared to be next to one another. Cannot believe it worked but
it looks great!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>