gpu analysis: add summary for similar/different
Here we see that there are differences in memory size, notably for Google
(only 16GB) and then the other clouds. On the other clouds (32GB) the actual
values are still a little different. The weirdest finding is the error correction
seems to be set at a mixture of yes/no for Azure GPUs

Signed-off-by: vsoch <vsoch@users.noreply.github.com>