analysis: mixbench cpu plots
These were really challenging to plot, because every run was
done differently and data files were interleaved. I started
with the amassed collection of csv files with the four
collections of data types (e.g., single/double/half/integer
precision) and then parsed from that the gflops and iops or
gflops, and then instead of line plots or scatter plots (they
also looked weird) I made a boxplot per size to show the
distribution of the gflops at a particular flops/byte being
used. I am not totally sure this is right but they look
reasonable so it is a good start.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>