plots: update colors and remove edges
This changeset updates all plot colors to have more distinct
values, and attempts to make them easier to see by removing the
edge borders. This also means, however, for plots with just
one value (that appeared as a flat line) we do not see anything.
This likely was not interesting data so I have removed the visuals
from readmes for now (but we still have the images) and then can
be added back or re-generated if needed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>