osu: fix runs for gpu 128 GKE and CE
We ran Google GPU setups with submit instead of run,
meaning it was not blocking, and they likely ran at the
same time and the result could be erroneous. Since we just
need to run for pairs we are doing a re-run for the study
paper size, 16 nodes or 128 GPU. We need to do the same
for Azure.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>